Revert ResourceTypeSelect
